validateMonth: implement TODO from pyinputplus.inputMonth
This commit aims to implement the changes to pysimplevalidate
that the pull request for pyinputplus
`inputMonth: add returnNumber and returnAbbreviation parameters`
expects.

validateMonth
- add parameters returnNumber and returnAbbreviation
- edit type signature for function to add two new bools
- in docstring, edit description of return value and type
- in docstring, edit typo 'an email address' -> 'a month'
- in docstring, describe returnNumber and returnAbbreviation
- in doctests, provide a case that uses returnNumber and a
    case that uses returnAbbreviation
- in case that input is an integer (the try/except)
-- give int(value) a name, monthAsIntegerValue
-- use that name
-- the `else` case contains what validateMonth returned before the
     changes in this commit
- in case `value[:3].upper() in monthNames.keys()`
-- the final `return` contains what validateMonth returned before
     the changes in this commit
-- the if/elif check for month_names == ENGLISH_MONTHS as the
     try/except does

I claim: for case `elif value.upper() in monthNames.values()`, the
body of that case it not currently reached.
Argument: monthNames.values() are not uppercase. so, value.upper()
will not occur in them.

Suppose we changed that value.upper() to value.title() to match
the case that is in monthNames. I claim control would still not
make it to the body, because any `value` that could get there
would first be caught by the `if value[:3].upper() in monthNames.keys()`.